Adapt things to the novel way we are handling networking
